Moved atomic fixnum classes into own files.
